fix: For added clarity move eos_library_helpers into just eos_helpers and delete the old files.
